Generate Python script within job submission dialog
