+++ app.html
@@ -50,6 +50,7 @@
 	<script type="text/javascript" src="app.js"></script>
 	<script type="text/javascript" src="model.js"></script>
 	<script type="text/javascript" src="controller.js"></script>
+	<script type="text/javascript" src="script_generator.js"></script>
 </head>
 <!-- Page passes the container for the graph to the program -->
 <!-- <body onload="main();">
